[VPlan, VP] 2/4 Introduce new live-in values and refactoring to support EVL recipe
This patch is based on the original work by Vineet Kumar: D104610

This patch introduces TripCount and RuntiVF as live-in VPValues whose
underlying IR Values are available only at plan execution time.
Canonical induction variable is cached so it can be reused when creating
EVL recipe.

This patch is second in the series of patches to enable predicated
vectorization support in LV introduced in RFC D99750.